Reenable type check to enable or disable module
According to the rdf:type of the selected resource the module will show off or
not. Therefore users can add URIs to the extension's config file in the same
way like the example entry for foaf:Document. This new behaviour is disabled by
default and can be enabled through a new config parameter.